fix: 프로젝트 알림 생성 버그 수정
