+++ src/service/project/project.ts
@@ -18,10 +18,3 @@ export async function getMyProjectDetail(projectId: string | bigint) {
     return await request('GET', `/api/project/detail?projectId=${projectId}`)
 }
 
-/**
- * 프로젝트 참여 요청
- * @param projectId
- */
-export async function requestParticipationProject(projectId: bigint, positionId: bigint) {
-    return await request('POST', `/api/project/participation?projectId=${projectId}`, { positionId });
-}
